Switch simple_types_filter to salt.utils
